+++ src/css/types.rs
@@ -1,0 +1,5 @@
+#[derive(Copy, Clone, PartialEq, Eq, Debug)]
+pub(crate) enum Importance {
+    Default,
+    Important,
+}